style(package): rename to helix-status
BREAKING CHANGE: Adjusts the package name to `helix-status`, as the library and service is no longer exclusive to Pingdom. The service will be available under a new name and the old package will be marked as deprecated on NPM with instructions to update to the new status library.

fixes #29 – update your `package.json` references from `@adobe/helix-pingdom-status` to `@adobe/helix-status`